+++ src/routes/topicRouter.ts
@@ -3,6 +3,8 @@ import express, {
     Response
 } from "express";
 import { Topic } from "../models";
+import { checkRingSignature, checkRingSignatureWithOwner,tokenAddress } from "../utils";
+
 const topicRouter = express.Router();
 
 // Get all topics
@@ -28,7 +30,9 @@ topicRouter.get("/:id", async (req: Request, res: Response) => {
 // Post a new topic
 topicRouter.post("/", async (req: Request, res: Response) => {
     try {
-        const topic = new Topic(req.body);
+        await checkRingSignature(req.body.ringSignature, tokenAddress);
+        // need to check how we get the owner keyImage, directly in the payload or from the irng signature
+        const topic = new Topic(req.body.topic);
         await topic.save();
         res.status(201).send(topic);
     } catch (error) {
@@ -39,7 +43,9 @@ topicRouter.post("/", async (req: Request, res: Response) => {
 // Delete a topic
 topicRouter.delete("/:id", async (req: Request, res: Response) => {
     try {
-        // TODO ADD CHECK ABOUT THE RINGSIGNATURE, whith the check about the senderKeyImage
+        // first we get the topic owner 
+        const topic = await Topic.findById(req.params.id);
+        await checkRingSignatureWithOwner(req.body.ringSignature, topic?.ownerkeyImage ?? "", tokenAddress);
         await Topic.findByIdAndDelete(req.params.id);
         res.status(200).send("Topic deleted");
     }
@@ -51,7 +57,8 @@ topicRouter.delete("/:id", async (req: Request, res: Response) => {
 // Modify a topic
 topicRouter.patch("/:id", async (req: Request, res: Response) => {
     try {
-        // TODO ADD CHECK ABOUT THE RINGSIGNATURE, whith the check about the senderKeyImage
+        const topic = await Topic.findById(req.params.id);
+        await checkRingSignatureWithOwner(req.body.ringSignature, topic?.ownerkeyImage ?? "", tokenAddress);
         await Topic.findByIdAndUpdate(req.params.id, req.body);
         await Topic.updateOne({
             _id: req.params.id
